LayoutTests/http/tests/security/credentials-from-different-domains-expected.txt

ALERT: Verifying credentials are used when given: User: user, password: pass.
ALERT: Verifying credentials are preemptively set when they have been given: User: user, password: pass.
ALERT: Verifying credentials are used after navigating within the same domain: User: user, password: pass.
ALERT: Verifying authentication fails from another domain: Authentication canceled
